+++ packages/components/src/layout/TotalLayout/TotalLayoutHeader/TotalLayoutHeader.constants.js
@@ -7,6 +7,10 @@ export const TOTAL_LAYOUT_HEADER_PROP_TYPES = {
   formTitlePlaceholder: PropTypes.string,
   compact: PropTypes.bool,
   icon: PropTypes.element,
+  onCancel: PropTypes.func,
+  children: PropTypes.node,
+  cancelable: PropTypes.bool,
+  direction: PropTypes.oneOf(['row', 'column']),
 };
 
 export const TOTAL_LAYOUT_HEADER_DEFAULT_PROPS = {
@@ -14,4 +18,6 @@ export const TOTAL_LAYOUT_HEADER_DEFAULT_PROPS = {
   icon: <Icon />,
   formTitlePlaceholder: '',
   compact: false,
+  cancelable: true,
+  direction: 'column',
 };