+++ packages/components/src/layout/VerticalStepperContainer/VerticalStepperContainer.stories.js
@@ -1,5 +1,6 @@
 import React, { useState, useMemo } from 'react';
 import PropTypes from 'prop-types';
+import { PluginAssignmentsIcon } from '@bubbles-ui/icons/solid';
 import { VerticalStepperContainer } from './VerticalStepperContainer';
 import { VERTICAL_STEPPER_CONTAINER_DEFAULT_PROPS } from './VerticalStepperContainer.constants';
 import mdx from './VerticalStepperContainer.mdx';
@@ -7,8 +8,14 @@ import { SIMPLE_DATA as STEPPER_DATA } from '../../navigation/VerticalStepper/mo
 import { Box } from '../Box';
 import { Stack } from '../Stack';
 import { ContextContainer } from '../ContextContainer';
-import { Button } from '../../form';
-import { Paragraph } from '../../typography';
+import { Button } from '../../form/Button';
+import { Paragraph } from '../../typography/Paragraph';
+import {
+  TotalLayoutFooterContainer,
+  TotalLayoutHeader,
+  TotalLayoutStepContainer,
+} from '../TotalLayout';
+import { TotalLayoutContainer } from '../TotalLayout/TotalLayoutContainer/TotalLayoutContainer';
 
 export default {
   title: 'Atoms/Layout/VerticalStepperContainer',
@@ -89,6 +96,114 @@ Template.propTypes = {
 export const Playground = Template.bind({});
 
 Playground.args = {
+  ...VERTICAL_STEPPER_CONTAINER_DEFAULT_PROPS,
+  data: STEPPER_DATA,
+};
+
+const TotalLayoutTemplate = ({ data = [], currentStep, ...props }) => {
+  const [activeStep, setActiveStep] = useState(currentStep || 0);
+  const scrollRef = React.useRef(null);
+
+  const randomContent = useMemo(
+    () => [...Array(data.length)].map(() => Math.floor(Math.random() * 10)),
+    [data],
+  );
+
+  // Prepare Header. It is necesary to pass the setOpenCancelModal function to the header
+  const handleOnCancel = () => {
+    console.log('Redirecting after cancel');
+  };
+
+  const handleNext = () => {
+    setActiveStep(activeStep + 1);
+    window.scrollTo(0, 0, { behavior: 'smooth' });
+  };
+
+  const handlePrev = () => {
+    setActiveStep(activeStep - 1);
+    window.scrollTo(0, 0, { behavior: 'smooth' });
+  };
+
+  const buildHeader = () => (
+    <TotalLayoutHeader
+      title={'Nueva Tarea'}
+      icon={<PluginAssignmentsIcon />}
+      formTitlePlaceholder={'Título de la tarea'}
+      onCancel={handleOnCancel}
+    />
+  );
+
+  return (
+    <Box style={{ margin: -16 }}>
+      <TotalLayoutContainer scrollRef={scrollRef} Header={() => buildHeader()}>
+        <VerticalStepperContainer
+          {...props}
+          data={data}
+          currentStep={activeStep}
+          scrollRef={scrollRef}
+        >
+          {
+            [...Array(data.length)].map((_, i) => (
+              <TotalLayoutStepContainer
+                key={`step-${i}`}
+                stepName={`Step ${i + 1}`}
+                Footer={
+                  <TotalLayoutFooterContainer
+                    noFlex
+                    fixed
+                    scrollRef={scrollRef}
+                    leftZone={
+                      <>
+                        {activeStep > 0 && (
+                          <Button variant="outline" onClick={handlePrev}>
+                            Previous
+                          </Button>
+                        )}
+                      </>
+                    }
+                    rightZone={<Button onClick={handleNext}>Next</Button>}
+                  />
+                }
+              >
+                <Box>
+                  {[...Array(randomContent[i])].map((__, j) => (
+                    <>
+                      <Paragraph key={`p1-${j}`}>
+                        Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
+                        tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,
+                        quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo
+                        consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse
+                        cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat
+                        non proident, sunt in culpa qui officia deserunt mollit anim id est laborum
+                      </Paragraph>
+                      <Paragraph key={`p2-${j}`}>
+                        Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
+                        tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,
+                        quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo
+                        consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse
+                        cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat
+                        non proident, sunt in culpa qui officia deserunt mollit anim id est laborum
+                      </Paragraph>
+                    </>
+                  ))}
+                </Box>
+              </TotalLayoutStepContainer>
+            ))[activeStep]
+          }
+        </VerticalStepperContainer>
+      </TotalLayoutContainer>
+    </Box>
+  );
+};
+
+TotalLayoutTemplate.propTypes = {
+  data: PropTypes.array,
+  currentStep: PropTypes.number,
+};
+
+export const WithTotalLayout = TotalLayoutTemplate.bind({});
+
+WithTotalLayout.args = {
   ...VERTICAL_STEPPER_CONTAINER_DEFAULT_PROPS,
   data: STEPPER_DATA,
 };