+++ packages/components/src/misc/FileIcon/AssetDocumentIcon/index.js
@@ -1,0 +1,2 @@
+export * from './AssetDocumentIcon';
+export * from './AssetDocumentIconSmall';