+++ packages/components/src/misc/index.js
@@ -7,5 +7,7 @@ export * from './FileIcon/AssetAudioIcon';
 export * from './FileIcon/AssetImageIcon';
 export * from './FileIcon/AssetBookmarkIcon';
 export * from './FileIcon/AssetVideoIcon';
+export * from './FileIcon/AssetMediaIcon';
+export * from './FileIcon/AssetDocumentIcon';
 export * from './AvatarSubject';
 export * from './CardEmptyCover';