+++ packages/editors/src/form/TextEditor/TextEditor.js
@@ -12,7 +12,7 @@ import Placeholder from '@tiptap/extension-placeholder';
 import Paragraph from '@tiptap/extension-paragraph';
 import { useExtensions } from '../../utils';
 import { BubbleMenu } from '../BubbleMenu';
-import { Toolbar, TOOLBAR_POSITIONS } from '../Toolbar';
+import { Toolbar, HeaderToolbar, TOOLBAR_POSITIONS } from '../Toolbar';
 import { TextEditorProvider } from '../TextEditorProvider';
 import { TextEditorStyles } from './TextEditor.styles';
 
@@ -39,6 +39,7 @@ export const TEXT_EDITOR_PROP_TYPES = {
   readOnly: PropTypes.bool,
   children: PropTypes.node,
   toolbarPortal: PropTypes.any,
+  scrollRef: PropTypes.any,
 };
 
 const TextEditor = ({
@@ -55,11 +56,13 @@ const TextEditor = ({
   acceptedTags = [],
   toolbarPosition,
   toolbarPortal,
+  scrollRef,
 }) => {
   const store = React.useRef({
     isFocus: false,
     acceptedTags: acceptedTags.join('|'),
   });
+  const [isToolbarReady, setIsToolbarReady] = React.useState(false);
   const extensions = useExtensions(children);
   const { classes, cx } = TextEditorStyles({}, { name: 'TextEditor' });
   const editor = useEditor({
@@ -198,9 +201,15 @@ const TextEditor = ({
     }
   }, [placeholder]);
 
+  React.useEffect(() => {
+    if (toolbarPortal) {
+      setIsToolbarReady(true);
+    }
+  }, [toolbarPortal]);
+
   const ToolbarComponent = React.useMemo(
     () =>
-      toolbarPortal
+      isToolbarReady
         ? () =>
             createPortal(
               <Toolbar
@@ -221,7 +230,7 @@ const TextEditor = ({
               {children}
             </Toolbar>
           ),
-    [toolbarPortal],
+    [isToolbarReady],
   );
 
   return (
@@ -235,11 +244,12 @@ const TextEditor = ({
       <TextEditorProvider editor={editor} readOnly={readOnly}>
         {readOnly ? null : (
           <Box style={{ zIndex: 1 }}>
-            <ToolbarComponent />
+            {isToolbarReady && <ToolbarComponent />}
             <BubbleMenu />
           </Box>
         )}
         <Box
+          ref={scrollRef}
           className={cx(classes.editorContainer, editorContainerClassname)}
           style={{ zIndex: 0 }}
         >