+++ packages/editors/CHANGELOG.md
@@ -1,5 +1,185 @@
 # @bubbles-ui/editors
 
+## 1.2.74
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.74
+  - @bubbles-ui/icons@1.2.74
+
+## 1.2.73
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.73
+  - @bubbles-ui/icons@1.2.73
+
+## 1.2.72
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.72
+  - @bubbles-ui/icons@1.2.72
+
+## 1.2.71
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.71
+  - @bubbles-ui/icons@1.2.71
+
+## 1.2.70
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.70
+  - @bubbles-ui/icons@1.2.70
+
+## 1.2.69
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.69
+  - @bubbles-ui/icons@1.2.69
+
+## 1.2.68
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.68
+  - @bubbles-ui/icons@1.2.68
+
+## 1.2.67
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.67
+  - @bubbles-ui/icons@1.2.67
+
+## 1.2.66
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.66
+  - @bubbles-ui/icons@1.2.66
+
+## 1.2.65
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.65
+  - @bubbles-ui/icons@1.2.65
+
+## 1.2.64
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.64
+  - @bubbles-ui/icons@1.2.64
+
+## 1.2.63
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.63
+  - @bubbles-ui/icons@1.2.63
+
+## 1.2.62
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.62
+  - @bubbles-ui/icons@1.2.62
+
+## 1.2.61
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.61
+  - @bubbles-ui/icons@1.2.61
+
+## 1.2.60
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.60
+  - @bubbles-ui/icons@1.2.60
+
+## 1.2.59
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.59
+  - @bubbles-ui/icons@1.2.59
+
+## 1.2.58
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.58
+  - @bubbles-ui/icons@1.2.58
+
+## 1.2.57
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.57
+  - @bubbles-ui/icons@1.2.57
+
+## 1.2.56
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.56
+  - @bubbles-ui/icons@1.2.56
+
+## 1.2.55
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.55
+  - @bubbles-ui/icons@1.2.55
+
 ## 1.2.54
 
 ### Patch Changes